requirements.txt
snowflake-snowpark-python==1.5.1
jupyter==1.0.0
python-dotenv==1.0.0